Move Schedulers provider to util/scheduler package
