+++ app/src/main/java/org/simple/clinic/allpatientsinfacility/AllPatientsInFacilityUiStateProducer.kt
@@ -8,6 +8,7 @@ import io.reactivex.rxkotlin.withLatestFrom
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.user.UserSession
+import org.simple.clinic.util.scheduler.SchedulersProvider
 import org.simple.clinic.widgets.ScreenCreated
 import org.simple.clinic.widgets.UiEvent
 import javax.inject.Inject
