+++ app/src/main/java/org/simple/clinic/util/scheduler/DefaultSchedulersProvider.kt
@@ -1,9 +1,8 @@
-package org.simple.clinic.util
+package org.simple.clinic.util.scheduler
 
 import io.reactivex.Scheduler
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import org.simple.clinic.allpatientsinfacility.SchedulersProvider
 
 class DefaultSchedulersProvider : SchedulersProvider {
   override fun io(): Scheduler = Schedulers.io()
