+++ app/src/test/java/org/simple/clinic/util/scheduler/TrampolineSchedulersProvider.kt
@@ -1,7 +1,8 @@
-package org.simple.clinic.allpatientsinfacility
+package org.simple.clinic.util.scheduler
 
 import io.reactivex.Scheduler
 import io.reactivex.schedulers.Schedulers
+import org.simple.clinic.util.scheduler.SchedulersProvider
 
 class TrampolineSchedulersProvider : SchedulersProvider {
   override fun io(): Scheduler = Schedulers.trampoline()
